+++ src/main/java/shootingstar/var/repository/User/UserRepositoryCustom.java
@@ -1,4 +1,4 @@
-package shootingstar.var.repository;
+package shootingstar.var.repository.User;
 
 public interface UserRepositoryCustom {
 }
