+++ src/main/java/shootingstar/var/repository/User/UserRepositoryCustomImpl.java
@@ -1,4 +1,4 @@
-package shootingstar.var.repository;
+package shootingstar.var.repository.User;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
